Add files via upload
Warby Parker is a transformative lifestyle brand with a lofty objective: to offer designer eyewear at a revolutionary price while leading the way for socially conscious businesses. Founded in 2010 and named after two characters in an early Jack Kerouac journal, Warby Parker believes in creative thinking, smart design, and doing good in the world — for every pair of eyeglasses and sunglasses sold, a pair is distributed to someone in need.

In this project, i will analyze different Warby Parker’s marketing funnels in order to calculate conversion rates. Here are the funnels and the tables that i am given:

Quiz Funnel:
- survey
Home Try-On Funnel:
- quiz
- home_try_on
- purchase